requirements.txt
flask
tensorflow==2.0.0
numpy
pillow